Revamped
Complete change to give readers a better understanding on how to make a virtual machine using Amazon Linux 2. This change was triggered after this from thread https://forums.aws.amazon.com/thread.jspa?messageID=828964&tstart=0, where I realized that the article was not changed despite the promise that someone would. My hope is this is a more clear and straight forward explanation.