LAMP Linux 2 edits for consistent nomenclature
Noticed that it was a bit odd that while user and group names were generally demarcated as inline-code-style items, only ec2-user was not. In addition, in a few places it was referred to by the somewhat unwieldy ec2-user user. By replacing these occurrences with inline-code-styled text, it not only makes the nomenclature more specific, but spares having to use the word user, because it clearly fits with the convention on the rest of the page of referencing such items (which appear in other command-line snippets already).