Removal of instructions for enhanced networking on Ubuntu as these are not supported by Canonical
Instructions removed for users to compile drivers from source. This method is not supported by
Canonical. Instead the user should install/upgrade the linux-aws kernel package if support is missing
or if driver is not up to date.

Canonical will continue to integrate new features and fixes for the ENA and ixgbevf driver in the
AWS-specific linux-aws kernel as part of our longstanding partnership.