Add info on when it is safe to resume I/O activity
Added an information on when to safely resume I/O activity when creating a snapshot for EBS volumes in a RAID array. The current documentation doesn't clearly say when it is safe to resume the I/O activity. Is it after the snapshot is complete? Or when the snapshot has been successfully initiated?

Based on this AWS re:Invent 2015 youtube video, it is safe to resume the I/O activity when the snapshot has been successfully initiated. 

Reference:
Source: https://www.youtube.com/watch?reload=9&v=2wKgha8CZ_w&feature=youtu.be&t=32m  
(See slide at 32 Minutes)