Clarification on Spot EC2 Partial Hour Billing
Hi Team,

I would like to request for clarification regarding the Spot Instance pricing. Say I have a Spot request with a maximum price of $0.04 which has been successfully fulfilled and the instance has already been running for 90 minutes (past the first instance hour). The Spot Price goes up to $0.05 which will automatically result in the interruption/termination of my instance. 

Will I be billed with $0.06 for that 90 minutes ($0.04 + $0.02), or will I only be paying $0.04, not counting the last 30 minutes?

Based on this document, including the official EC2 FAQ, the answer should be $0.04 since the instance was interrupted in the subsequent hour. In the "Specifying a Duration for Your Spot Instances" section, there is a paragraph that says: 

"You are billed at this price for each hour or partial hour that the instance is running. A partial instance hour is billed to the nearest second."

The official EC2 FAQ also says that: 

Q. How will I be charged if my Spot instance is interrupted?

If your Spot instance is terminated or stopped by Amazon EC2 in the first instance hour, you will not be charged for that usage. However, if you terminate the instance yourself, you will be charged to the nearest second. If the Spot instance is terminated or stopped by Amazon EC2 in any subsequent hour, you will be charged for your usage to the nearest second. If you are running on Windows and you terminate the instance yourself, you will be charged for an entire hour.

Reference: https://aws.amazon.com/ec2/faqs/


However, there seems to be a conflicting document that says otherwise. In the "How Spot Instances Work" document, it says:

"If you terminate your instance, you pay for any partial hour used (as you do for On-Demand or Reserved Instances). However, you are not charged for any partial hour of usage if the Spot price goes above your maximum price and Amazon EC2 interrupts your Spot Instance."

https://docs.aws.amazon.com/whitepapers/latest/cost-optimization-leveraging-ec2-spot-instances/how-spot-instances-work.html

This is where the confusion is. It says that you are not charged for any partial hour of usage, which contradicts the first two documents above. Hence, I would like to ask for clarification regarding this specific statement from the above paragraph:

"However, you are not charged for any partial hour of usage if the Spot price goes above your maximum price and Amazon EC2 interrupts your Spot Instance."

In relation to the example that I mentioned above, does the phrase: "partial hour of usage" only refers to the first instance hour, and not the subsequent hour? This document is quite ambiguous and personally, it seems that this should be updated to avoid any confusion to the readers. 

Thank you in advance for your assistance. Looking forward to your reply.

Cheers,
Jon Bonso