Updated examples, to use JMESpath queries
JMESpath is built into the AWS CLI (https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-output.html#cli-usage-output-filter). jq is a separate binary that would need to be installed to make examples that pipe JSON to jq work, thus it seems it would be a better approach to use the features that are built into the cli, in order to filter output.