Add ssh option to explicitly used named key
The ssh command will attempt to connect to the instance by offering each of the ssh keys registered with the ssh agent on the user's system and then offering the key provided explicitly on the command-line from the '-i' option.  Arguably this order seems wrong, but that is the behavior of openssh.  For users that have a number of keys registered with their agent the host may reject the connection attempt due to too many failed keys (default for this is 5) without ever trying the key specified on the command-line.

This patch adds the `IdentitiesOnly` ssh option which has ssh ignore the keys registered with the agent and only attempt a connection using the key specified by the '-i' option.  This ensures that the correct key is used, and only that key.  It provides a smoother, more consistent experience and helps with the support experience as failures with this command-line eliminate some additional debugging of the key exchange.