Include disclaimer about cron expression syntax
Added a disclaimer about the modified syntax of the cron expression used in Spot Requests when compared to the linked documentation for cron schedules in Amazon CloudWatch.
The CloudWatch documentation states that the cron scheduling only goes to the granularity of *minutes*, while Spot Requests cron expressions support granularity to the *second*. This can cause errors for users when they attempt to schedule scaling actions via cron expressions, as a cron expression correctly formatted to the CloudWatch documentation linked here will fail on the validation in the Spot Request as the fields will all be off by one (e.g. the user has the Day of Week data Mon-Fri in what is the Month field for Spot Requests, and has the year field absent).