Update AmazonEBS.md
I think you should consider making the first sentence on line #22 singular. Doing so makes it more consistent with how most of the AWS docs are written where the singular is used, and it fits better with the follow up sentence "To make a volume available..." which is also singular.  When the sentence is plural "EBS volumes are created in a specific Availability Zone", it can be interpreted that all EMS volumes are created in a particular Availability Zone.  To be fair I don't think most human readers would interpret it this way, but I think the change is still worth considering. Thanks for reviewing the proposal.