Update ebs-volume-types.md
GP2 & GP3 SSDs are not just for "development and test environments". These types of SSDs can also be used in non-critical production workloads which don't require high IOPS performance. If you add the line "development and test environments" then people might wrongly think that GP2 or GP3 cannot be used in a production environment, which is quite wrong. GP2 or GP3 can be used in production workloads that don't require high IOPS performance.