Clarify support for legacy address space
RIPE now supports ROA/RPKI for legacy space. See https://www.ripe.net/manage-ips-and-asns/resource-management/rpki/resource-certification-rpki-for-provider-independent-end-users. 
With that the previous statement that LEGACY space is not supported is no longer true. As long as one can generate a valid ROA for that LEGACY address space it will be supported.